+++ Assets/Scripts/Models/Signal/LyricDisplaySignal.cs
@@ -3,10 +3,10 @@ using System.Threading.Tasks;
 
 namespace InnerMediaPlayer.Models.Signal
 {
-    internal struct LyricDisplaySignal : ITaskQueueBinding<int>
+    internal struct LyricDisplaySignal : ITaskQueueBinding<long>
     {
-        public int Param1 { get; set; }
-        public Func<int, Tools.CancellationTokenSource, IProgress<TaskStatus>, Task> Func { get; set; }
+        public long Param1 { get; set; }
+        public Func<long, Tools.CancellationTokenSource, IProgress<TaskStatus>, Task> Func { get; set; }
         public Action CallBack { get; set; }
     }
 }
